added player class, human inherites from player methods move, initialize, and define_symbol; computer inherited from player define_symbol method
