+++ lib/player.rb
@@ -1,0 +1,24 @@
+# This class definies player behavior
+class Player
+  def initialize(userinteraction, markers, name)
+    @ui = userinteraction
+    @markers = markers
+    @name = name
+  end
+
+  def move(board, symbol)
+    spot = set_spot
+    until board.valid(board.get_value(spot))
+      @ui.space_taken
+      spot = set_spot
+    end
+    board.insert_value(spot, symbol) if board.valid(board.get_value(spot))
+    # @ui.great_move
+  end
+
+  def define_symbol(hum)
+    return @markers.o if hum == @markers.x
+
+    @markers.x
+  end
+end
